requirements.txt
ultralytics[export]
selenium
python-dotenv
fastapi[standard]